dhcp4client: default to math.Rand over crypto.Rand
This prevents blocking until the CRNG is initialized with Linux ≥ 4.16:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=43838a23a05fbd13e47d750d3dfd77001536dd33

As per https://tools.ietf.org/html/rfc2131#section-4.1, the xid field should be
chosen such a way as to minimize the chance of using an xid identical to one
used by another client.

This only requires a seeded random number generator, not a cryptographically
secure random number generator.